+++ src/app/components/vio/button/VioButtonColored.vue
@@ -29,19 +29,19 @@
 </template>
 
 <script setup lang="ts">
-interface Props {
+const {
+  ariaLabel,
+  disabled = false,
+  isPrimary = true,
+  to = undefined,
+  type = 'button',
+} = defineProps<{
   ariaLabel: string
   disabled?: boolean
   isPrimary?: boolean
   to?: string
   type?: 'button' | 'reset' | 'submit'
-}
-withDefaults(defineProps<Props>(), {
-  disabled: false,
-  isPrimary: true,
-  to: undefined,
-  type: 'button',
-})
+}>()
 
 const emit = defineEmits<{
   click: []