+++ src/playground/nuxt.config.ts
@@ -37,6 +37,9 @@ export default defineNuxtConfig(
       site: {
         twitter: '@dargmuesli',
       },
+      sitemap: {
+        zeroRuntime: true,
+      },
     },
     VIO_NUXT_BASE_CONFIG({
       siteName: SITE_NAME,