+++ src/shared/utils/constants.ts
@@ -6,9 +6,6 @@ export { IS_IN_FRONTEND_DEVELOPMENT } from '../../node/static'
 
 export const VIO_SITE_NAME = 'Vio'
 
-export const SITE_URL =
-  process.env.NUXT_PUBLIC_I18N_BASE_URL ||
-  `https://${process.env.HOST || 'localhost'}:${process.env.PORT || '3000'}`
 export const CACHE_VERSION = 'zeMtipb6C9'
 export const COOKIE_CONTROL_CONSENT_COOKIE_NAME =
   DEFAULTS.cookieNameIsConsentGiven