+++ tests/playwright.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig, devices } from '@playwright/test'
 
-import { SITE_URL } from './e2e/utils/constants'
+import { SITE_URL, TIMEOUT } from '#tests/e2e/utils/constants'
 
 /**
  * Read environment variables from file.
@@ -86,12 +86,12 @@ export default defineConfig({
 
   /* Run your local dev server before starting the tests */
   webServer: {
-    command: `pnpm run --dir ../src start:${process.env.VIO_SERVER || 'dev'}`,
+    command: `pnpm run --dir ../src certificates && pnpm run --dir ../src start:${process.env.VIO_SERVER || 'dev'}`,
     env: {
       NUXT_PUBLIC_VIO_IS_TESTING: 'true',
     },
     ignoreHTTPSErrors: true, // TODO: remove once tests run without it
-    timeout: process.env.NODE_ENV === 'production' ? 10000 : 100000,
+    timeout: TIMEOUT,
     url: SITE_URL,
     reuseExistingServer: !process.env.CI,
   },